requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
cloudinary==1.22.0
cryptography==3.2.1
Django==3.1.3
django-filter==2.4.0
django-rest-jwt==0.1.46
django-rest-knox==4.1.0
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Markdown==3.3.3
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python-http-client==3.3.1
pytz==2020.4
sendgrid==6.4.7
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.26.1
